Separate ProofSender on its own module
